+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - DQL query execution can now be started from the editor with a shortcut `ctrl` + `enter`. If the user selected a part
   of the query or the cursor is inside a subquery (for example, `append` command), IDE will as the user to decide which
   part of the query they want to execute.
+- Chart tooltips now display the X-axis value as a bold title above the series data row inside tooltips.
 
 ### Bug fixes